rename folder 'lang' --> 'ctl-lang' for easier inclusion
