+++ CMakeLists.txt
@@ -42,10 +42,11 @@ target_include_directories(${PROJECT_NAME} PUBLIC
         ${yahashcombine_SOURCE_DIR}/include
         ${argvparse_SOURCE_DIR}/include
         ${CMAKE_CURRENT_BINARY_DIR}
+        src/ctl-lang
         include
         src)
 
-add_subdirectory(src/lang)
+add_subdirectory(src/ctl-lang)
 target_link_libraries(${PROJECT_NAME} ctl_lang)
 add_executable(${PROJECT_NAME}_demo src/main.cpp)
 target_link_libraries(${PROJECT_NAME}_demo ${PROJECT_NAME} argvparse)
